integer overflow conversionを修正

+++ packages/backend/app/usecases/users/ports/get_params.go
@@ -4,12 +4,12 @@ package ports
 type FollowsGetParams struct {
 	PreferredUserID           string
 	PreferredUserPaginationID string
-	Limit                     uint64
+	Limit                     int
 }
 
 // FollowersGetParams is the parameters to get follows.
 type FollowersGetParams struct {
 	PreferredUserID           string
 	PreferredUserPaginationID string
-	Limit                     uint64
+	Limit                     int
 }
